Sped up the kernel operator a bit more. ~1 sec on my home desktop

+++ fyp2016/Ascan.h
@@ -4,15 +4,15 @@
 class Ascan
 {
 private:
-	int offset;
-	const int length;
+	unsigned int offset;
+	const unsigned int length;
 	const uint16_t *values;
 public:
-	Ascan(int, uint16_t*);
+	Ascan(unsigned int, uint16_t*);
 	~Ascan();
 
-	int getSize();
-	int getIndex(int);
-	int getPercent(float);
+	unsigned int getSize();
+	uint16_t getIndex(unsigned int);
+	uint16_t getPercent(float);
 };
